+++ demo/tc-ui/browser/lib/relationship-editor.js
