+++ demo/tc-ui/browser/pages/view.js
