+++ demo/tc-ui/delete.js
