+++ demo/tc-ui/lib/configure-schema.js
