+++ demo/tc-ui/templates/components/buttons.jsx
