+++ demo/tc-ui/templates/components/edit-components/edit-boolean.jsx
