+++ demo/tc-ui/templates/components/edit-components/edit-relationships.jsx
