+++ demo/tc-ui/templates/components/edit-components/edit-text.jsx
