+++ demo/tc-ui/templates/components/edit-components/index.jsx
