+++ demo/tc-ui/templates/layout/asset-loading.jsx
