+++ demo/tc-ui/templates/view-page.jsx
