+++ package.json
@@ -5,17 +5,25 @@
   "description": "A set of tools for building adnmin interfaces on top of graph data",
   "main": "index.js",
   "dependencies": {
+    "@financial-times/lambda-logger": "^2.0.4",
     "@financial-times/tc-api-db-manager": "file:./packages/tc-api-db-manager",
     "@financial-times/tc-api-rest-handlers": "file:./packages/tc-api-rest-handlers",
     "@financial-times/tc-schema-publisher": "file:./packages/tc-schema-publisher",
-    "@financial-times/tc-schema-sdk": "file:./packages/tc-schema-sdk",
+    "@financial-times/tc-schema-sdk": "file:packages/tc-schema-sdk",
+    "autolinker": "^3.11.1",
+    "awesomplete": "^1.1.5",
     "aws-sdk": "^2.504.0",
     "body-parser": "^1.19.0",
+    "date-fns": "^2.8.1",
     "express": "^4.17.1",
     "graphql-playground-middleware-express": "^1.7.12",
+    "http-errors": "^1.7.3",
     "neo4j-driver": "^1.7.6",
     "node-fetch": "^2.6.0",
-    "preact": "^10.0.5"
+    "preact": "^10.0.5",
+    "preact-render-to-string": "^5.1.2",
+    "qs": "^6.9.1",
+    "showdown": "^1.9.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.0",
