--- packages/tc-ui/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@financial-times/tc-ui",
-  "version": "0.0.0",
-  "description": "from biz-ops admin/runbooks",
-  "main": "index.js",
-  "author": "",
-  "license": "ISC",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Financial-Times/treecreeper.git"
-  },
-  "dependencies": {
-    "@financial-times/lambda-logger": "^2.0.4",
-    "@financial-times/tc-schema-sdk": "file:../tc-schema-sdk",
-    "autolinker": "^3.11.1",
-    "awesomplete": "^1.1.5",
-    "date-fns": "^2.8.1",
-    "http-errors": "^1.7.3",
-    "node-fetch": "^2.6.0",
-    "preact": "^10.0.5",
-    "preact-render-to-string": "^5.1.2",
-    "qs": "^6.9.1",
-    "showdown": "^1.9.1"
-  },
-  "devDependencies": {}
-}
